feat(tsos-compiler): emit all compiler diagnostics
I discovered that there 2 types of diagnostics, those we get from the compiler and those that we get from the emitter.

Both types or diagnostics are now returned together. I don't feel the need to make a point of difference between them, to me an error or warning is just that and it should be fixed regardless of what stage of the compilation process caused it.

BREAKING CHANGE: Some scripts that ran fine before may now return diagnostic errors.